position, techStack 엔티티 global 패키지로 이전
