+++ src/main/java/com/eum/post/model/entity/PostPosition.java
@@ -1,5 +1,6 @@
 package com.eum.post.model.entity;
 
+import com.eum.global.model.entity.Position;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
