+++ src/main/java/com/eum/post/model/repository/PositionRepository.java
@@ -1,6 +1,6 @@
 package com.eum.post.model.repository;
 
-import com.eum.post.model.entity.Position;
+import com.eum.global.model.entity.Position;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
