+++ src/main/java/com/eum/post/model/repository/TechStackRepository.java
@@ -1,6 +1,6 @@
 package com.eum.post.model.repository;
 
-import com.eum.post.model.entity.TechStack;
+import com.eum.global.model.entity.TechStack;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
